leaf_classifier_deployment/requirements.txt
torch>=2.0.0,<2.5.0
torchvision>=0.15.0,<0.20.0
Pillow>=9.0.0,<11.0.0